Fix signature

+++ src/link.c
@@ -123,7 +123,7 @@ static void CB2_LinkTest(void);
 static void ProcessRecvCmds(u8 unused);
 static void LinkCB_SendHeldKeys(void);
 static void ResetBlockSend(void);
-static bool32 InitBlockSend(const void *src, size_t size);
+static bool32 InitBlockSend(const void *src, u32 size);
 static void LinkCB_BlockSendBegin(void);
 static void LinkCB_BlockSend(void);
 static void LinkCB_BlockSendEnd(void);
